functions/pub_sub_relay/requirements.txt
google-cloud-storage
google-cloud-kms == 1.0.0
google-cloud-pubsub
six==1.12.0
smmap2==2.0.5
mock==3.0.5
